feat: adding sUSD to synths

+++ contracts/Interfaces/curve/Curve.sol
@@ -44,9 +44,9 @@ interface ICurveFi {
         bool use_underlying
     ) external payable;
 
-    function coins(uint256) external returns (address);
+    function coins(int128) external returns (address);
 
-    function underlying_coins(uint256) external returns (address);
+    function underlying_coins(int128) external returns (address);
 
     function remove_liquidity_imbalance(
         uint256[2] calldata amounts,
